+++ UI/TitlebarTabs/TitlebarTabsController.swift
@@ -118,6 +118,7 @@ final class TitlebarTabsController: NSViewController {
 
     @objc
     private func handleDocumentStoreDidChange(_ notification: Notification) {
+        guard notification.isOnlySidebarChromeChange == false else { return }
         rebuildTabs()
     }